requirements.txt
mplfinance
reportlab
statsmodels
python-dotenv
oandapyV20
requests
